backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.28.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydoe2==1.3.0
allpairspy==2.5.0
pandas==2.1.2
redis==5.0.1
sentry-sdk==1.33.1
prometheus-client==0.17.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1 